password strength display implemented
